EventSystem

Documentation

handleKeys :: Event -> GameState -> GameState

Is responsible for communication betwen program and user, performes an action for each legal input signal

canChangeTo :: Orientation -> GameState -> Bool

Chcecks if given move can be performed for current GameState or if orientation wasn't alredy changed in considered frame