GameSettings

Documentation

fps :: Int

Frames generated per secound - sets pace of the game (Can be modified)

sizeOfTile :: Int

Given in pixels (Can be modified)

sizeOfBoard :: Int

Given in tiles (Can be modified)

sizeOfScreen :: Int

Computes size of the window in pixels

rewardForEatingApple :: Int

Sets an amount of tiles the snake grows each time it eats an apple (Can be modified)

bgColor :: Color

Color of background outside of the map (Visible only in fullscreen-mode) (Can be modified)

snakeColor :: Color

Color of snake's body (Can be modified)

appleColor :: Color

Color of apples (Can be modified)

wallColor :: Color

Color of walls (Can be modified)

gridColor :: Color

Color of grid in the background of the map, half of the tiles will be a little bit lighter (Can be modified)