UpdateGame

Documentation

moveSnake :: GameState -> ([Position], Int)

Simulate moving of snake, returns snakeBody and tilesToBeAdded after motion

ateApple :: [Position] -> Position

Function is invoked after eating a apple, returns newApplePos

hasCollisionOccured :: [Position] -> Bool

If returns true, game is finnished

update :: Float -> GameState -> GameState

Function is a single step between one state to another